added module for hash, sign, and verify
