+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod wallet;
 pub mod commands;
 pub mod error;
+pub mod file_ops;
